requirements.txt
discord.py==1.6.0
pymongo==3.11.2
pyyaml==5.4.1
pygal==2.4.0
pillow==8.3.0